+++ ydb/library/workload/tpcc/transaction_payment.cpp
@@ -287,7 +287,7 @@ NThreading::TFuture<TStatus> GetPaymentTask(
 
     const auto& in = FixedTransactionInputs<TInputs>(context, [&] {
         TInputs generated;
-        generated.WarehouseID = context.WarehouseID;
+        generated.WarehouseID = static_cast<int>(context.WarehouseID);
         generated.DistrictID = RandomNumber(DISTRICT_LOW_ID, DISTRICT_HIGH_ID);
         generated.PaymentAmount = static_cast<double>(RandomNumber(100, 500000)) / 100.0;
